chore(ellipsis): split demo
