+++ src/packages/__VUE/ellipsis/doc.en-US.md
@@ -14,83 +14,25 @@ const app = createApp();
 app.use(Ellipsis);
 ```
 
-### Leading
+### Start
 
-:::demo
+> demo: ellipsis start
 
-```vue
-<template>
-  <nut-ellipsis
-    content="NutUI3.0上线后我们研发团队也在不断的优化、测试、使用、迭代 Vue3 的相关组件，但是在跨端小程序的开发过程中，发现没有合适的组件库可以支持多端开发。为了填补这一空白，同时为了优化开发者体验，让 NutUI 能够为更多的开发者带来便利，我们决定在 NutUI 中增加小程序多端适配的能力。"
-    direction="start"
-  ></nut-ellipsis>
-</template>
-```
+### End
 
-:::
-
-### Tailing
-
-:::demo
-
-```vue
-<template>
-  <nut-ellipsis
-    content="NutUI3.0上线后我们研发团队也在不断的优化、测试、使用、迭代 Vue3 的相关组件，但是在跨端小程序的开发过程中，发现没有合适的组件库可以支持多端开发。为了填补这一空白，同时为了优化开发者体验，让 NutUI 能够为更多的开发者带来便利，我们决定在 NutUI 中增加小程序多端适配的能力。"
-    direction="end"
-  ></nut-ellipsis>
-</template>
-```
-
-:::
+> demo: ellipsis end
 
 ### Middle
 
-:::demo
-
-```vue
-<template>
-  <nut-ellipsis
-    content="NutUI3.0上线后我们研发团队也在不断的优化、测试、使用、迭代 Vue3 的相关组件，但是在跨端小程序的开发过程中，发现没有合适的组件库可以支持多端开发。为了填补这一空白，同时为了优化开发者体验，让 NutUI 能够为更多的开发者带来便利，我们决定在 NutUI 中增加小程序多端适配的能力。"
-    direction="middle"
-  ></nut-ellipsis>
-</template>
-```
-
-:::
+> demo: ellipsis middle
 
 ### Multi-line
 
-:::demo
-
-```vue
-<template>
-  <nut-ellipsis
-    content="NutUI3.0上线后我们研发团队也在不断的优化、测试、使用、迭代 Vue3 的相关组件，但是在跨端小程序的开发过程中，发现没有合适的组件库可以支持多端开发。为了填补这一空白，同时为了优化开发者体验，让 NutUI 能够为更多的开发者带来便利，我们决定在 NutUI 中增加小程序多端适配的能力。"
-    direction="start"
-    rows="3"
-  ></nut-ellipsis>
-</template>
-```
-
-:::
+> demo: ellipsis row
 
 ### Expand & Collapse
 
-:::demo
-
-```vue
-<template>
-  <nut-ellipsis
-    direction="start"
-    expandText="Expand"
-    collapseText="Collapse"
-    content="NutUI3.0上线后我们研发团队也在不断的优化、测试、使用、迭代 Vue3 的相关组件，但是在跨端小程序的开发过程中，发现没有合适的组件库可以支持多端开发。为了填补这一空白，同时为了优化开发者体验，让 NutUI 能够为更多的开发者带来便利，我们决定在 NutUI 中增加小程序多端适配的能力。"
-  ></nut-ellipsis>
-</template>
-```
-
-:::
+> demo: ellipsis expand
 
 ## API
