+++ src/packages/__VUE/ellipsis/doc.md
@@ -16,81 +16,23 @@ app.use(Ellipsis);
 
 ### 头部省略
 
-:::demo
-
-```vue
-<template>
-  <nut-ellipsis
-    content="NutUI3.0上线后我们研发团队也在不断的优化、测试、使用、迭代 Vue3 的相关组件，但是在跨端小程序的开发过程中，发现没有合适的组件库可以支持多端开发。为了填补这一空白，同时为了优化开发者体验，让 NutUI 能够为更多的开发者带来便利，我们决定在 NutUI 中增加小程序多端适配的能力。"
-    direction="start"
-  ></nut-ellipsis>
-</template>
-```
-
-:::
+> demo: ellipsis start
 
 ### 尾部省略
 
-:::demo
-
-```vue
-<template>
-  <nut-ellipsis
-    content="NutUI3.0上线后我们研发团队也在不断的优化、测试、使用、迭代 Vue3 的相关组件，但是在跨端小程序的开发过程中，发现没有合适的组件库可以支持多端开发。为了填补这一空白，同时为了优化开发者体验，让 NutUI 能够为更多的开发者带来便利，我们决定在 NutUI 中增加小程序多端适配的能力。"
-    direction="end"
-  ></nut-ellipsis>
-</template>
-```
-
-:::
+> demo: ellipsis end
 
 ### 中间省略
 
-:::demo
-
-```vue
-<template>
-  <nut-ellipsis
-    content="NutUI3.0上线后我们研发团队也在不断的优化、测试、使用、迭代 Vue3 的相关组件，但是在跨端小程序的开发过程中，发现没有合适的组件库可以支持多端开发。为了填补这一空白，同时为了优化开发者体验，让 NutUI 能够为更多的开发者带来便利，我们决定在 NutUI 中增加小程序多端适配的能力。"
-    direction="middle"
-  ></nut-ellipsis>
-</template>
-```
-
-:::
+> demo: ellipsis middle
 
 ### 多行省略
 
-:::demo
-
-```vue
-<template>
-  <nut-ellipsis
-    content="NutUI3.0上线后我们研发团队也在不断的优化、测试、使用、迭代 Vue3 的相关组件，但是在跨端小程序的开发过程中，发现没有合适的组件库可以支持多端开发。为了填补这一空白，同时为了优化开发者体验，让 NutUI 能够为更多的开发者带来便利，我们决定在 NutUI 中增加小程序多端适配的能力。"
-    direction="start"
-    rows="3"
-  ></nut-ellipsis>
-</template>
-```
-
-:::
+> demo: ellipsis row
 
 ### 展开收起
 
-:::demo
-
-```vue
-<template>
-  <nut-ellipsis
-    direction="start"
-    expandText="展开"
-    collapseText="收起"
-    content="NutUI3.0上线后我们研发团队也在不断的优化、测试、使用、迭代 Vue3 的相关组件，但是在跨端小程序的开发过程中，发现没有合适的组件库可以支持多端开发。为了填补这一空白，同时为了优化开发者体验，让 NutUI 能够为更多的开发者带来便利，我们决定在 NutUI 中增加小程序多端适配的能力。"
-  ></nut-ellipsis>
-</template>
-```
-
-:::
+> demo: ellipsis expand
 
 ## API
